requirements.txt
accelerate==1.7.0
apex==0.9.10dev
attrdict==2.0.1
av==14.4.0
bert_score==0.3.13
bitsandbytes==0.49.2
blobfile==3.2.0
cog==0.17.2
data_processing==0.1.4
datasets==3.6.0
decord==0.6.0
deepspeed==0.18.9
diffusers==0.37.1
editdistance==0.8.1
einops==0.8.2
einops_exts==0.0.4
fastapi==0.135.3
fastchat==0.1.0
ffmpeg==1.4
fire==0.7.1
flash_attn==2.7.4.post1
ftfy==6.3.1
gradio==6.12.0
green_score==0.0.12
horovod==0.28.1
huggingface_hub==0.32.4
imageio==2.37.3
librosa==0.11.0
mathruler==0.1.0
matplotlib==3.10.8
nibabel==5.4.2
nltk==3.9.1
numpy==2.4.4
open_flamingo==2.0.1
opencv_python_headless==4.11.0.86
packaging==26.0
pandas==3.0.2
Pillow==12.2.0
openai==1.93.0
pycocoevalcap==1.2
PyYAML==6.0.3
qwen_vl_utils==0.0.14
RaTEScore==0.6.0
ray==2.46.0
regex==2024.11.6
Requests==2.33.1
rich==15.0.0
ring_flash_attn==0.1.8
rouge==1.0.1
safetensors==0.7.0
scikit_learn==1.8.0
seaborn==0.13.2
setuptools==78.1.1
sglang==0.5.10.post1
shortuuid==1.0.13
soundfile==0.13.1
taming_transformers==0.0.1
tenacity==9.1.2
timm==1.0.26
tokenizers==0.22.2
torch==2.7.0
torch_npu==2.9.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.15.0
tyro==1.0.12
uvicorn==0.44.0
vector_quantize_pytorch==1.28.1
vllm==0.9.0.1
vocos==0.1.0
wandb==0.25.1
xformers==0.0.30
